Added Player Class
