+++ player.py
@@ -1,0 +1,38 @@
+from __future__ import division
+
+import sys
+import mmath
+
+if sys.version_info[0] >= 3:
+    xrange = range
+
+class Player():
+    def __init__(self, *args, **kwargs):
+        return
+    
+    def hit_test(self, position, world, vector, max_distance=8):
+        """ Line of sight search from current position. If a block is
+        intersected it is returned, along with the block previously in the line
+        of sight. If no block is found, return None, None.
+
+        Parameters
+        ----------
+        position : tuple of len 3
+            The (x, y, z) position to check visibility from.
+        vector : tuple of len 3
+            The line of sight vector.
+        max_distance : int
+            How many blocks away to search for a hit.
+
+        """
+        m = 8
+        x, y, z = position
+        dx, dy, dz = vector
+        previous = None
+        for _ in xrange(max_distance * m):
+            key = mmath.normalize((x, y, z))
+            if key != previous and key in world:
+                return key, previous
+            previous = key
+            x, y, z = x + dx / m, y + dy / m, z + dz / m
+        return None, None
